requirements.txt
@angular/cli
@angular/compiler-cli
django
django-heroku
django_rest_framework
facebook-py
krakenex
psycopg2
psycopg2-binary
python-twitter
requests
watson_developer_cloud
